mm-video: qdsp6: Bumping up Q6 clock to max for interlaced contents.
Bumping Q6 clock to maximum for interlaced contents.
A few interlaced content was not playing with normal
Q6 clock so it needs be bumped up. The cause of this
issue is maibly high instantaneous bitrate which can
occur in non-interlaced content too. Making this
change for interlaced content for now. It needs
more refinement to bump up Q6 clock only for
high bitrate clips.

Change-Id: Id4b45c54416c276394951ac59012cb0a7ea5f581